Add a first draft of the ExamineArgs trigger
This trigger allows a trigger to examine the arguments of the
interposed function, if the arguments are strings or integers.

One or more arguments may be skipped using the <skip> tag.
It may be used more than once, in order to ignore several
function arguments.

The <argType>, <argCompare>, and <argBaseArrayChooser> tags
must be specified exactly once, to examine a function argument
and perhaps trigger true:

* `argType`: "string" | "int"
* `argCompare`: "strstr" | "strcmp" | "equal" | "and"
    * For "string" type, use either "strstr" or "strcmp"
    * For "int" type, use either "equal" or "and"
* `argBaseArrayChooser`: an integer greater than or equal to zero.

The `libfi -E /path/to/c/code` flag must be used to specify a path
to a C/C++ file that defines the following arrays:

    char **examine_args_string[];
    int examine_args_string_lastindex[];
    int *examine_args_int[];
    int examine_args_int_lastindex[];

These four arrays define the one-or-more values that the trigger
will use the `argCompare` comparison function to trigger.  The
`argBaseArrayChooser` specifies which element of the `examine_args_string`
or `examine_args_int` array to use.
The `examine_args_string_lastindex` and `examine_args_int_lastindex`
arrays specify how many values to compare.

For an example, see these links:
* https://github.com/basho/eleveldb/blob/d49d5de0c42af00af1f22be2cca01db4a0cfa3d0/priv/scenario/intercept_commonpaths.xml
* https://github.com/basho/eleveldb/blob/d49d5de0c42af00af1f22be2cca01db4a0cfa3d0/priv/scenario/intercept_commonpaths.cpp